# backend/docs/openapi/openapi.yaml
...
openapi: 3.1.0
paths:
...
  /companies/keycloak/{keycloakId}:
    get:
      summary: Buscar empresa por ID do Keycloak
      tags:
      - Companies
      parameters:
      - name: keycloakId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
...
  /customers:
    get:
      summary: Listar todos os clientes
      tags:
      - Customers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
    post:
      summary: Criar novo cliente
      tags:
      - Customers
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequestDTO"
        required: true
      responses:
        "201":
          description: Cliente criado com sucesso
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - admin
        - company
  /customers/company/{keycloakId}:
    get:
      summary: Buscar clientes por ID da empresa
      tags:
      - Customers
      parameters:
      - name: keycloakId
        in: path
        required: true
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Clientes encontrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResultCustomerResponseDTO"
        "404":
          description: Empresa não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
  /customers/keycloak/{keycloakId}:
    get:
      summary: Buscar cliente por ID do Keycloak
      tags:
      - Customers
      parameters:
      - name: keycloakId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Cliente encontrado
        "404":
          description: Cliente não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - client
  /customers/paged:
    get:
      summary: Listar clientes paginados
      tags:
      - Customers
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResultCustomerResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
  /customers/search:
    get:
      summary: "Buscar clientes por nome, CPF ou e-mail com paginação"
      tags:
      - Customers
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: query
        in: query
        schema:
          type: string
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResultCustomerResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
  /customers/{id}:
    put:
      summary: Atualizar cliente
      tags:
      - Customers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequestDTO"
        required: true
      responses:
        "200":
          description: Cliente atualizado com sucesso
        "404":
          description: Cliente não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - admin
        - company
    get:
      summary: Buscar cliente por ID
      tags:
      - Customers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Cliente encontrado
        "404":
          description: Cliente não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
    delete:
      summary: Excluir cliente
      tags:
      - Customers
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Cliente removido com sucesso
        "404":
          description: Cliente não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
...
  /users/role/{role}:
    get:
      parameters:
      - name: role
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
      summary: Get Users By Role
      tags:
      - Keycloak User Resource
  /users/with-role/{role}:
    post:
      parameters:
      - name: role
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Create User With Role
      tags:
      - Keycloak User Resource
  /users/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Update User
      tags:
      - Keycloak User Resource
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
      summary: Delete User
      tags:
      - Keycloak User Resource
components:
  schemas:
...
    CustomerRequestDTO:
      type: object
      required:
      - name
      - cpf
      - email
      - phone
      - companyId
      - occupation
      - street
      - streetNumber
      - city
      - neighborhood
      - state
      - zipCode
      - country
      properties:
        name:
          type: string
          pattern: \S
        cpf:
          type: string
          pattern: "^(\\d{3}\\.\\d{3}\\.\\d{3}-\\d{2}|\\d{11})$"
          minLength: 11
          maxLength: 14
        email:
          type: string
          pattern: \S
        phone:
          type: string
          pattern: "^\\+?[0-9]{10,15}$"
        birthDate:
          $ref: "#/components/schemas/LocalDate"
        companyId:
          type: integer
          format: int64
        occupation:
          type: string
          pattern: \S
        street:
          type: string
          pattern: \S
        streetNumber:
          type: number
          minimum: 1
        city:
          type: string
          pattern: \S
        neighborhood:
          type: string
          pattern: \S
        state:
          type: string
          pattern: \S
        zipCode:
          type: string
          pattern: \S
        country:
          type: string
          pattern: \S
        keycloakId:
          type: string
        createdBy:
          type: string
        updatedBy:
          type: string
        createdAt:
          $ref: "#/components/schemas/LocalDate"
        updatedAt:
          $ref: "#/components/schemas/LocalDate"
        salary:
          type: number
          exclusiveMinimum: 0.0
    CustomerResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        keycloakId:
          type: string
        enabled:
          type: boolean
        name:
          type: string
        cpf:
          type: string
        email:
          type: string
        phone:
          type: string
        birthDate:
          $ref: "#/components/schemas/LocalDate"
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        occupation:
          type: string
        street:
          type: string
        streetNumber:
          type: number
        city:
          type: string
        neighborhood:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
        createdBy:
          type: string
        updatedBy:
          type: string
        createdAt:
          $ref: "#/components/schemas/LocalDate"
        updatedAt:
          $ref: "#/components/schemas/LocalDate"
        virtualAccountNumber:
          type: string
        virtualAgencyNumber:
          type: string
        virtualBalance:
          type: number
        salary:
          type: number
...
    LocalDate:
      type: string
      format: date
      examples:
      - 2022-03-10
...
    PagedResultCustomerResponseDTO:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CustomerResponseDTO"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8080/realms/crediflow/.well-known/openid-configuration
      description: Authentication
...
info:
  title: backend API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8082
  description: Auto generated value
- url: http://0.0.0.0:8082
  description: Auto generated value